requirements.txt
streamlit==1.18.0
altair==4.2.2
matplotlib==3.7.1
numpy==1.26.4
Pillow==9.5.0
scikit-learn==1.3.0
seaborn==0.13.2
tensorflow==2.10.1
gdown==5.2.0